+++ app/src/main/java/com/almikey/jiplace/model/Geolocation.kt
@@ -1,4 +1,4 @@
 package com.almikey.jiplace.model
 
 
-class MyLocation(var longitude:Float=0.toFloat(),var latitude:Float=0.toFloat())
+class MyLocation(var longitude:Double = 0.toDouble(),var latitude:Double=0.toDouble())
